+++ actionview/lib/action_view/helpers/form_options_helper.rb
@@ -353,24 +353,56 @@ module ActionView
       #   # => <option value="Advanced">Advanced</option>
       #   # => <option value="Super Platinum" disabled="disabled">Super Platinum</option>
       #
+      # If you wish to specify priority option tags, so that they will
+      # be listed above the rest of the (long) list set +selected+ to be a hash, with <tt>:priority</tt> being
+      # either a value or array of values to be placed at the top of the options. You can then also specify
+      # how the separator should look like using the <tt>:priority_separator</tt> option, which can be a String
+      # or a Proc receiving the container. Set the <tt>:priority_unique</tt> option to +true+ if the values
+      # in the priority list should be unique, i.e. they will not be shown below the separator.
+      #
+      #   options_for_select(["Afghanistan", "Bahrain", "Cook Islands", "USA"], priority: ["USA"])
+      #   # => <option value="USA">USA</option>
+      #   # => <option value="" disabled="disabled">-------------</option>
+      #   # => <option value="Afghanistan">Afghanistan</option>
+      #   # => <option value="Bahrain">Bahrain</option>
+      #   # => <option value="Cook Islands">Cook Islands</option>
+      #   # => <option value="USA">USA</option>
+      #
       # NOTE: Only the option tags are returned, you have to wrap this call in a regular HTML select tag.
       def options_for_select(container, selected = nil)
         return container if String === container
 
-        selected, disabled = extract_selected_and_disabled(selected).map do |r|
-          Array(r).map(&:to_s)
-        end
+        selected, disabled, priority, priority_separator, priority_unique = extract_select_options(selected)
+        disabled = Array(disabled).map(&:to_s)
+        priority_options, options = [], []
 
-        container.map do |element|
+        container.each do |element|
           html_attributes = option_html_attributes(element)
           text, value = option_text_and_value(element).map(&:to_s)
 
-          html_attributes[:selected] ||= option_value_selected?(value, selected)
-          html_attributes[:disabled] ||= disabled && option_value_selected?(value, disabled)
+          html_attributes[:selected] ||= selected.include?(value)
+          html_attributes[:disabled] ||= disabled && disabled.include?(value)
           html_attributes[:value] = value
 
-          tag_builder.content_tag_string(:option, text, html_attributes)
-        end.join("\n").html_safe
+          option = tag_builder.content_tag_string(:option, text, html_attributes)
+
+          if priority.include?(value)
+            priority_options << option
+
+            unless priority_unique
+              option = tag_builder.content_tag_string(:option, text, html_attributes.except!(:selected))
+              options << option
+            end
+          else
+            options << option
+          end
+        end
+
+        return options.join("\n").html_safe if priority_options.empty?
+
+        priority_separator = priority_separator.is_a?(Proc) ? priority_separator.call(container) : priority_separator
+        priority_separator_option = tag_builder.content_tag_string(:option, priority_separator, value: "", disabled: true)
+        (priority_options + [priority_separator_option] + options).join("\n").html_safe
       end
 
       # Returns a string of option tags that have been compiled by iterating over the +collection+ and assigning
@@ -389,9 +421,10 @@ module ActionView
       # If +selected+ is specified as a Proc, those members of the collection that return true for the anonymous
       # function are the selected values.
       #
-      # +selected+ can also be a hash, specifying both <tt>:selected</tt> and/or <tt>:disabled</tt> values as required.
+      # +selected+ can also be a hash, specifying <tt>:selected</tt>, <tt>:disabled</tt>, <tt>:priority</tt>,
+      # <tt>:priority_separator</tt> and/or <tt>:priority_unique</tt> values as required.
       #
-      # Be sure to specify the same class as the +value_method+ when specifying selected or disabled options.
+      # Be sure to specify the same class as the +value_method+ when specifying selected, disabled or priority options.
       # Failure to do this will produce undesired results. Example:
       #   options_from_collection_for_select(@people, 'id', 'name', '1')
       # Will not select a person with the id of 1 because 1 (an Integer) is not the same as '1' (a string)
@@ -401,10 +434,14 @@ module ActionView
         options = collection.map do |element|
           [value_for_collection(element, text_method), value_for_collection(element, value_method), option_html_attributes(element)]
         end
-        selected, disabled = extract_selected_and_disabled(selected)
+        selected, disabled, priority, priority_separator, priority_unique = extract_select_options(selected)
+
         select_deselect = {
           selected: extract_values_from_collection(collection, value_method, selected),
-          disabled: extract_values_from_collection(collection, value_method, disabled)
+          disabled: extract_values_from_collection(collection, value_method, disabled),
+          priority: priority,
+          priority_separator: priority_separator,
+          priority_unique: priority_unique
         }
 
         options_for_select(options, select_deselect)
@@ -573,24 +610,16 @@ module ActionView
       # NOTE: Only the option tags are returned, you have to wrap this call in
       # a regular HTML select tag.
       def time_zone_options_for_select(selected = nil, priority_zones = nil, model = ::ActiveSupport::TimeZone)
-        zone_options = "".html_safe
-
         zones = model.all
-        convert_zones = lambda { |list| list.map { |z| [ z.to_s, z.name ] } }
 
         if priority_zones
           if priority_zones.is_a?(Regexp)
             priority_zones = zones.select { |z| z =~ priority_zones }
           end
-
-          zone_options.safe_concat options_for_select(convert_zones[priority_zones], selected)
-          zone_options.safe_concat content_tag("option".freeze, "-------------", value: "", disabled: true)
-          zone_options.safe_concat "\n"
-
-          zones = zones - priority_zones
+          priority_zones = priority_zones.map(&:name)
         end
 
-        zone_options.safe_concat options_for_select(convert_zones[zones], selected)
+        options_from_collection_for_select(zones, :name, :to_s, selected: selected, priority: priority_zones, priority_unique: true)
       end
 
       # Returns radio button tags for the collection of existing return values
@@ -776,18 +805,20 @@ module ActionView
           end
         end
 
-        def option_value_selected?(value, selected)
-          Array(selected).include? value
-        end
-
-        def extract_selected_and_disabled(selected)
+        def extract_select_options(selected)
           if selected.is_a?(Proc)
-            [selected, nil]
+            [selected, nil, nil, nil, nil]
           else
             selected = Array.wrap(selected)
             options = selected.extract_options!.symbolize_keys
             selected_items = options.fetch(:selected, selected)
-            [selected_items, options[:disabled]]
+            [
+              Array(selected_items).map(&:to_s),
+              options[:disabled],
+              Array(options[:priority]).map(&:to_s),
+              options.fetch(:priority_separator) { "-------------" },
+              options.fetch(:priority_unique, false)
+            ]
           end
         end