+++ actionview/lib/action_view/helpers/tags/select.rb
@@ -14,10 +14,8 @@ module ActionView
         end
 
         def render
-          option_tags_options = {
-            selected: @options.fetch(:selected) { value },
-            disabled: @options[:disabled]
-          }
+          option_tags_options = @options.dup
+          option_tags_options[:selected] = @options.fetch(:selected) { value }
 
           option_tags = if grouped_choices?
             grouped_options_for_select(@choices, option_tags_options)